[ISSofaPython] Add search method bindings to BaseContext to be able to retrieve objects in the scene graph from their type
The code supersedes the `getObjects` binding method which was never used
in practice.
The semantic used for the methods is `search[All]Dir`,
e.g `SearchAllUp`, `SearchLocal`.
Search methods work with object types, described by a class name and an
optional template name.
It contrasts with the `getObject` method which relies on a path to
retrieve an object instance, which requires by definition the
knowledge of the name of the object we are looking for to succeed.

Ref T6299

(cherry picked from commit 64e5cf2a54900106bb67428572358df94a5770ba)