[ISSofaPython] UPDATE: virtual env: also use the system site-packages
The virtualenv creation option --system-site-packages is set by default,
but options can be changed by the user.

This allows to use some python packages already installed on the system
(pip does not install a dependency in the virtual env if it finds the
suitable version in the system site-packages).

The virtual env is re-created if the associated options were modified
by the user.

//Ref// T6507

(cherry picked from commit 37f3f335566e30af1c131626b6433965c4dc6a01)