[ISSofaPython] UPDATE: 'activate' the virtual env, setting the Python home to the venv dir
ISSOFAPYTHON_VIRTUALENV_DIR is now properly taken into account
and we try to set the Python home to the virtual env dir before
initializing the interpreter.
This is enough to 'activate' the virtual env, and in my tests it
was more complete than manually adding the underlying site-packages
dir.

In H3D the same thing should be done to really use the virtual env.
(to simplify we could maybe remove the possibility to customize the
virtual env dir, and always set it to <build-dir>/python-venv/)

//Ref// T6507

(cherry picked from commit aa07c76b48abb5fd1e7a79995615216a2ef25012)