[ISSofaPython] UPDATE: embed a standalone Python when possible
We copy the standalone Python that was recently added in ISExternals,
into the build tree as well as into the installed binaries dir (only on Windows).
This embedded distribution of Python is used for the build
(see the CMakeLists.txt in ISExternals/python2) and the runtime.
This is to progress on making the software easily copyable and
runnable on another computer.

When a standalone Python is not available (so currently on Linux), we
use a regular virtual env, which stays linked to the Python installed
on the system.

Regarding the C++ code, whether we are using a real virtualenv or
a fully standalone python is transparent
(in all cases, there is just to set the python home to the root dir
of the python env, before initializing the interpreter)

(//Ref// T6562)

(cherry picked from commit fc57cccb949e6d34c0e123273756fd98689292cb)