[ISSofaPython] Use fallback de-serialization of type using istream operator
This is done only when de-serializing from a string, and that the
underlying type we are initializing is not a string type.
In order for this to work, it relies entirely about having a
proper implementation of the istream operator for the underlying type.
This fallback mode allows for example to initialise and array of scalars
from an array of char, each encoding a scalar value.

//Ref// T6787

(cherry picked from commit 63cce5fe9eea6743a713f5cb53b63f2e285ba057)