[ISSofaPython] Try pybind11 cast method from python to c++ type when assigning a Data value
This is at least required when assigning boolean values.
If numpy is imported, it is possible that boolean types on the python
side get to be represented using numpy.bool_ instead of typical python bool.
With the previous implementation, this would throw an invalid assignement
exception since numpy.bool_ are not represented on the python with the
pybind11::bool_ type ( and before that the assignment would be silently
ignored, which I believe why the error is caught only now ).
However numpy bool can still be cast as c++ bool types using pybind11
cast method, which is what the current implementation does.

//Ref// T8180

(cherry picked from commit b653d86e9634e6f71459eb855eb98d96bff71004)