[ISSofaPython/build] CHANGE python sitecustomize.py: ISSOFAPYTHON_ADDITIONAL_SYS_PATH_DIRS CMake var allows to add project dirs to sys.path
This var is empty by default and can be set by a specific project
for eg. from ProjectConfiguration.cmake.
It allows to have the python sys.path correctly set up for the project,
without having to modify it in python launcher scripts or things like
that.
This should work in all cases, that is either when using the embedded python
of the build tree, the one of the install tree, and the one of the
packaged binaries.

However, what we do the most is still to install our python packages
with pip + setup.py files. The only thing to know is that in such case
the python package is entirely copied inside the libs dir of the
packaged python env.

(cherry picked from commit 38379e281721b402444a03b6d199d8f2d4564a7b)