[ISSofaPython]  FIX: Base-derived python bindings must now use pybind11::multiple_inheritance() (//Ref// T10602)
Issues (exceptions or crashes or missing attributes) appear in pybind11 for classes with virtual parents if the pybind11::multiple_inheritance() flag is not used.

This is necessary because with a virtual parent the base and derived classes do not have the same address in memory.
In previous versions, prior to 6283e640b, both the direct parent as well as Base were set as base classes, enabling internally the multiple_inheritance logic within pybind11.
Note that this lead to another issue with sofa::sptr casting, which is fixed in the follow-up commit.

(cherry picked from commit 2a473e9c9c184ee9e22dfe1f52d363bb1686257f)