[pybind11]  FIX: enable implicit casts for boost::intrusive_ptr-like custom holders and bindings with multiple_inheritance enabled (//Ref// T10602)
This fix rely on the existance of a holder constructor from a raw pointer, which is available for boost::intrusive_ptr but may have a different semantic for other cases.
So it may not be safe to apply upstream as-is.

(cherry picked from commit 6f6337face196d7254c6083b8cb8f728f6cbd253)